Correct inconsistent behaviour on PAMleft searches
revcom() does its work in two stages:
1. complement the sequence (taking advantage of the fact that the DNA
bases are encoded so that they are binary complements of each other)
2. reverse it (two bits at a time).

However while the complementing explicitly takes care to flip the PAM
bit, the reversing drops the results. This is not a problem where
PAM_right=1, as it is flipped to 0 and thus ignoring it is the same as
paying attention to it. But if PAM_right=0 it is flipped and then
ignored, causing inconsistent results.